// old-congress/shared/enums/needy-type.enum.ts
export enum NeedyType {
  Student = "אברך",
  Young = "בחור ישיבה",
  Welfare = "נתמך מחלקת רווחה",
  Needy = "נתמך קמחא דפסחא",
  BereavedFamily = "משפחות חרבות ברזל",
  Evacuee = "מפונים",
  Rabbi = "רב קהילה",
  Gabbai = "גבאי",
}